+++ R/print.R
@@ -37,43 +37,41 @@
 #'
 #' @export
 print.IssueTB <- function(x, ...) {
-    issue <- x
-
     issue_desc <- paste0(
         "Issue ",
-        issue[["owner"]],
+        x[["owner"]],
         "/",
-        issue[["repo"]],
+        x[["repo"]],
         "#",
-        issue[["number"]]
+        x[["number"]]
     )
 
     cli::cli_h2(
         cli::style_hyperlink(
             text = issue_desc,
-            url = issue[["html_url"]]
+            url = x[["html_url"]]
         )
     )
 
     cat(
         crayon::underline("Title:"),
         " ",
-        substr(x = issue[["title"]], start = 1L, stop = 80L),
+        substr(x = x[["title"]], start = 1L, stop = 80L),
         "\n",
         crayon::underline("Text:\n"),
         ifelse(
-            test = nchar(issue[["body"]]) > 320L,
+            test = nchar(x[["body"]]) > 320L,
             yes = paste0(
-                substr(x = issue[["body"]], start = 1L, stop = 320L),
+                substr(x = x[["body"]], start = 1L, stop = 320L),
                 "\n...\n"
             ),
-            no = issue[["body"]]
+            no = x[["body"]]
         ),
         "\n",
         sep = ""
     )
 
-    return(invisible(issue))
+    return(invisible(x))
 }
 
 #' @rdname print
@@ -81,7 +79,7 @@ print.IssueTB <- function(x, ...) {
 #' @method print IssuesTB
 #' @export
 print.IssuesTB <- function(x, ...) {
-    cat(crayon::bold(
+    crayon::bold(
         switch(
             EXPR = as.character(nrow(x)),
             "0" = "No issues",
@@ -89,7 +87,7 @@ print.IssuesTB <- function(x, ...) {
             paste("There are", nrow(x), "issues.")
         ),
         "\n"
-    ))
+    ) |> cat()
     for (id_issue in seq_len(nrow(x))) {
         cat("\n")
         print(x[id_issue, , drop = TRUE])
@@ -113,23 +111,23 @@ print.summary.IssueTB <- function(x, ...) {
         if (is.null(x$labels_color)) {
             cat(x$labels, sep = ", ")
         } else {
-        cat(
-            vapply(
-                X = seq_along(x$labels),
-                FUN = function(k) {
-                    label_style <- crayon::combine_styles(
-                        crayon::make_style(x$labels_color[k]),
-                        crayon::make_style(x$labels_bgcolor[k], bg = TRUE)
-                    )
-                    cli::style_hyperlink(
-                        text = label_style(x$labels_name[k]),
-                        url = x$labels_url[k]
-                    )
-                },
-                FUN.VALUE = character(1L)
-            ),
-            sep = ", "
-        )
+            cat(
+                vapply(
+                    X = seq_along(x$labels),
+                    FUN = function(k) {
+                        label_style <- crayon::combine_styles(
+                            crayon::make_style(x$labels_color[k]),
+                            crayon::make_style(x$labels_bgcolor[k], bg = TRUE)
+                        )
+                        cli::style_hyperlink(
+                            text = label_style(x$labels_name[k]),
+                            url = x$labels_url[k]
+                        )
+                    },
+                    FUN.VALUE = character(1L)
+                ),
+                sep = ", "
+            )
         }
 
         cat("\n")
@@ -176,6 +174,7 @@ print.summary.IssueTB <- function(x, ...) {
             "\n"
         )
     }
+    return(invisible(x))
 }
 
 #' @rdname print
@@ -185,13 +184,12 @@ print.summary.IssueTB <- function(x, ...) {
 print.summary.IssuesTB <- function(x, ...) {
     cat(
         crayon::bold(
-            if (x$nbr_issues == 0L) {
-                "No issues"
-            } else if (x$nbr_issues == 1L) {
-                "There is 1 issue: "
-            } else {
-                paste("There are", x$nbr_issues, "issues:")
-            }
+            switch(
+                EXPR = as.character(x$nbr_issues),
+                "0" = "No issues",
+                "1" = "There is 1 issue.",
+                paste("There are", x$nbr_issues, "issues.")
+            )
         ),
         paste0(
             "\n- ",
@@ -204,14 +202,72 @@ print.summary.IssuesTB <- function(x, ...) {
         ),
         "\n"
     )
+    return(invisible(x))
 }
 
 #' @rdname print
-#' @exportS3Method print summary.LabelsTB
-#' @method print summary.LabelsTB
+#' @exportS3Method print LabelsTB
+#' @method print LabelsTB
 #' @export
 print.LabelsTB <- function(x, ...) {
-    message("To document and to write ! #TODO")
+
+    x$labels_bgcolor <- x$color
+    x$labels_color <- c("grey8", "ivory")[
+        isDark(x$labels_bgcolor) + 1L
+    ]
+    x$labels_url <- paste(
+        "https://github.com",
+        x$owner,
+        x$repo,
+        "labels",
+        utils::URLencode(x$labels_name),
+        sep = "/"
+    )
+    x$formated_label <- vapply(
+        X = seq_len(nrow(x)),
+        FUN = function(k) {
+            label_style <- crayon::combine_styles(
+                crayon::make_style(x$labels_color[k]),
+                crayon::make_style(x$labels_bgcolor[k], bg = TRUE)
+            )
+            cli::style_hyperlink(
+                text = label_style(x$name[k]),
+                url = x$labels_url[k]
+            )
+        },
+        FUN.VALUE = character(1L)
+    )
+
+    couples <- unique(x[, c("owner", "repo")])
+    crayon::bold(
+        switch(
+            EXPR = as.character(nrow(couples)),
+            "0" = "No labels",
+            "1" = "There is 1 repo.",
+            paste("There are", nrow(couples), "repos.")
+        )
+    ) |> cat()
+
+    for (id in seq_len(nrow(couples))) {
+        owner_name <- couples[id, "owner"]
+        repo_name <- couples[id, "repo"]
+        labels_owner <- x |> subset(owner == owner_name & repo == repo_name)
+
+        cat(
+            paste0(
+                "\n- ",
+                cli::style_hyperlink(
+                    text = paste(owner_name, repo_name, sep = "/"),
+                    url = paste("https://github.com", owner_name, repo_name, sep = "/")
+                ),
+                ":"
+            ),
+            labels_owner$formated_label,
+            sep = ", "
+        )
+    }
+    cat("\n")
+    return(invisible(x))
 }
 
 #' @rdname print
@@ -219,5 +275,36 @@ print.LabelsTB <- function(x, ...) {
 #' @method print summary.LabelsTB
 #' @export
 print.summary.LabelsTB <- function(x, ...) {
-    message("To document and to write ! #TODO")
+    # print(unclass(x))
+    # return(NULL)
+    crayon::bold(
+        switch(
+            EXPR = as.character(nrow(x)),
+            "0" = "No labels",
+            "1" = "There is 1 label.",
+            paste("There are", nrow(x), "labels.")
+        )
+    ) |> cat()
+
+    couples <- unique(x[, c("owner", "repo")])
+
+    for (id in seq_len(nrow(couples))) {
+        owner_name <- couples[id, "owner"]
+        repo_name <- couples[id, "repo"]
+        labels_owner <- x |> subset(owner == owner_name & repo == repo_name)
+
+        cat(
+            paste0(
+                "\n- ",
+                cli::style_hyperlink(
+                    text = paste(owner_name, repo_name, sep = "/"),
+                    url = paste("https://github.com", owner_name, repo_name, sep = "/")
+                ),
+                ":"
+            ),
+            paste0(labels_owner$formated_label, ": ", labels_owner$description),
+            sep = "\n    - "
+        )
+    }
+    return(invisible(x))
 }